Make sure compatibility macros are defined before use
PKG_CHECK_VAR is only available in >=pkg-config-0.28. To allow the
configure script to be executed on a distribution with a lower
pkg-config version (e.g. CentOS 7) it is also defined locally.
However so far it is defined after being used, so it would result
in an error such as:

  configure.ac:61: error: possibly undefined macro: PKG_CHECK_VAR

Signed-off-by: Reto Gantenbein <reto.gantenbein@linuxmonk.ch>